+++ doc/macro-examples/increment.py
@@ -1,0 +1,1 @@
+increment('tests/test1')